Правка багов к уроку "Введение в базы данных"

+++ src/main/java/ru/hogwarts/school/controller/FacultyController.java
@@ -18,32 +18,43 @@ public class FacultyController
         this.facultyService = facultyService;
     }
 
-    @PostMapping("")
+    @PostMapping
     public ResponseEntity<Faculty> createFaculty(@RequestBody Faculty faculty)
     {
+        if(faculty == null)
+            return ResponseEntity.notFound().build();
         return ResponseEntity.ok(facultyService.addFaculty(faculty));
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Faculty> getStudent(@PathVariable Long id)
     {
-        return ResponseEntity.ok(facultyService.getFaculty(id));
+        Faculty tmpFaculty = facultyService.getFaculty(id);
+        if(tmpFaculty == null)
+            return ResponseEntity.notFound().build();
+        return ResponseEntity.ok(tmpFaculty);
     }
 
     @DeleteMapping("{id}")
     public ResponseEntity<Faculty> deleteStudent(@PathVariable Long id)
     {
+        Faculty tmpFaculty = facultyService.getFaculty(id);
+        if(tmpFaculty == null)
+            return ResponseEntity.notFound().build();
         return ResponseEntity.ok(facultyService.removeFaculty(id));
     }
 
-    @PutMapping()
+    @PutMapping
     public ResponseEntity<Faculty> updateStudent(@RequestBody Faculty faculty)
     {
+        if(faculty == null || facultyService.getFaculty(faculty.getId()) == null)
+            return ResponseEntity.notFound().build();
+
         return ResponseEntity.ok(facultyService.changeFaculty(faculty));
     }
 
-    @GetMapping("/sortByColor/{color}")
-    public List<Faculty> sortByColor(@PathVariable String color)
+    @GetMapping("/sortByColor")
+    public List<Faculty> sortByColor(@RequestParam(name = "color") String color)
     {
         return facultyService.sortByColor(color);
     }
